backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
build==1.2.2.post1
certifi==2025.1.31
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
filelock==3.18.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
dj-database-url
pyproject_hooks==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
setuptools==79.0.0
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.30.0
django-allauth
dj-rest-auth
PyJWT[crypto]
requests
cloudinary
django-cloudinary-storage
Pillow
django-filter
